requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
asyncpg==0.30.0
azure-storage-blob==12.23.0
azure-storage-queue==12.12.0
aiohttp==3.10.5
structlog==24.4.0
tenacity==9.0.0
python-multipart==0.0.12
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest-cov==5.0.0
httpx==0.27.0
